access system

+++ app/controllers/ListaController.php
@@ -35,32 +35,54 @@ class ListaController
 
     public function delete(array $params)
     {
-        $tipo=$params[0];
-        $table=($tipo=='funcionarios')?'users':'students';
-        $model=new UsersModel();
-        if($table=='users')
+        if($_SESSION['access']=='admin')
         {
-            $message=$model->deleteById($table,$params[1],$_SESSION['user']);
+            $tipo=$params[0];
+            $table=($tipo=='funcionarios')?'users':'students';
+            $model=new UsersModel();
+            if($table=='users')
+            {
+                $message=$model->deleteById($table,$params[1]);
+            }else
+            {
+                $message=$model->deleteById($table,$params[1]);
+            }
+            echo"<script>
+            alert('$message')
+            setTimeout(window.location.href='/lista/$tipo',2000)
+            </script>";
         }else
         {
-            $message=$model->deleteById($table,$params[1]);
+            echo "<script>
+            alert('Você não tem permissão para realizar essa ação')
+            setTimeout(window.location.href='/lista/funcionarios',2000)
+            </script>
+            ";   
         }
-        echo"<script>
-        alert('$message')
-        setTimeout(window.location.href='/lista/$tipo',2000)
-        </script>";
     }
 
     public function edit(array $params)
     {
-        $model=new UsersModel();
-        $table=($params[0]=='funcionarios')?'users':'students';
-        view('editar',[
-            'title'=>'editar',
-            'tipo'=>$params[0],
-            'id'=>$params[1],
-            'values'=>$model->selectById($table,$params[1])
-        ]);
+        if($_SESSION['access']=='admin')
+        {
+            $model=new UsersModel();
+            $setValues=$model->getAccessSet();
+            $table=($params[0]=='funcionarios')?'users':'students';
+            view('editar',[
+                'title'=>'editar',
+                'tipo'=>$params[0],
+                'id'=>$params[1],
+                'values'=>$model->selectById($table,$params[1]),
+                'setValues'=>$setValues
+            ]);
+        }else
+        {
+            echo "<script>
+            alert('Você não tem permissão para realizar essa ação')
+            setTimeout(window.location.href='/lista/funcionarios',2000)
+            </script>
+            ";
+        }
     }
 
     public function save(array $params)
